pyproject.toml
[project]
name = "financial-control"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
]
